enable Emacs silent rules during build
The "silent" rules make the build process of Emacs much less verbose.  This, in turn, makes it much easier to discover errors in the process when installing or upgrading with the brew `--verbose` option.  Since the intent is to allow humans to diagnose why the build process is going wrong – and the full commands are still logged by the build process – remove the unconditional disable.